+++ build-less-to-css.sh
@@ -5,4 +5,5 @@ lessc ./build.less ../css/theme.css --clean-css="--s1 --advanced" --source-map=.
 cd ../../clean-contrasted/less
 lessc ./build.less ../css/theme.css --clean-css="--s1 --advanced" --source-map=../css/theme.css.map
 cd ../../clean-bordered/less
+lessc ./build.less ../css/theme.css --clean-css="--s1 --advanced" --source-map=../css/theme.css.map
 cd ../../..