+++ resources/js/humhub.clean.theme.js
@@ -1,8 +1,8 @@
 humhub.module('ui.theme', function (module, require, $) {
     module.export({
         getContentTop: function () {
-            var $topBar = $('#topbar');
-            return $topBar.offset().top + $topBar.height() - $(window).scrollTop();
+            const topBar = $('#topbar');
+            return topBar.offset().top + topBar.height() - $(window).scrollTop();
         }
     });
 });